fix: missing dependency and per reviewer comments
