+++ samples/snapshot/pom.xml
@@ -30,6 +30,11 @@
       <artifactId>google-cloud-video-intelligence</artifactId>
       <version>1.5.5</version>
     </dependency>
+    <dependency>
+      <groupId>com.google.cloud</groupId>
+      <artifactId>google-cloud-storage</artifactId>
+      <version>1.113.4</version>
+    </dependency>
 
     <dependency>
       <groupId>junit</groupId>
