+++ samples/snippets/src/main/java/video/DetectPerson.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.example.video;
+package video;
 
-// [START video_detect_person_gcs]
+// [START video_detect_person]
 
 import com.google.api.gax.longrunning.OperationFuture;
 import com.google.cloud.videointelligence.v1.AnnotateVideoProgress;
@@ -33,21 +33,27 @@ import com.google.cloud.videointelligence.v1.VideoAnnotationResults;
 import com.google.cloud.videointelligence.v1.VideoContext;
 import com.google.cloud.videointelligence.v1.VideoIntelligenceServiceClient;
 import com.google.cloud.videointelligence.v1.VideoSegment;
+import com.google.protobuf.ByteString;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
-public class DetectPersonGcs {
+public class DetectPerson {
 
-  public static void detectPersonGcs() throws Exception {
+  public static void detectPerson() throws Exception {
     // TODO(developer): Replace these variables before running the sample.
-    String gcsUri = "gs://cloud-samples-data/video/googlework_short.mp4";
-    detectPersonGcs(gcsUri);
+    String localFilePath = "resources/googlework_short.mp4";
+    detectPerson(localFilePath);
   }
 
-  // Detects people in a video stored in Google Cloud Storage using
-  // the Cloud Video Intelligence API.
-  public static void detectPersonGcs(String gcsUri) throws Exception {
+  // Detects people in a video stored in a local file using the Cloud Video Intelligence API.
+  public static void detectPerson(String localFilePath) throws Exception {
     try (VideoIntelligenceServiceClient videoIntelligenceServiceClient =
         VideoIntelligenceServiceClient.create()) {
       // Reads a local video file and converts it to base64.
+      Path path = Paths.get(localFilePath);
+      byte[] data = Files.readAllBytes(path);
+      ByteString inputContent = ByteString.copyFrom(data);
 
       PersonDetectionConfig personDetectionConfig =
           PersonDetectionConfig.newBuilder()
@@ -61,18 +67,20 @@ public class DetectPersonGcs {
 
       AnnotateVideoRequest request =
           AnnotateVideoRequest.newBuilder()
-              .setInputUri(gcsUri)
+              .setInputContent(inputContent)
               .addFeatures(Feature.PERSON_DETECTION)
               .setVideoContext(videoContext)
               .build();
 
       // Detects people in a video
+      // We get the first result because only one video is processed.
       OperationFuture<AnnotateVideoResponse, AnnotateVideoProgress> future =
           videoIntelligenceServiceClient.annotateVideoAsync(request);
 
       System.out.println("Waiting for operation to complete...");
       AnnotateVideoResponse response = future.get();
-      // Get the first response, since we sent only one video.
+
+      // Gets annotations for video
       VideoAnnotationResults annotationResult = response.getAnnotationResultsList().get(0);
 
       // Annotations for list of people detected, tracked and recognized in video.
@@ -110,4 +118,4 @@ public class DetectPersonGcs {
     }
   }
 }
-// [END video_detect_person_gcs]
+// [END video_detect_person]
