+++ samples/snippets/src/test/java/com/example/video/DetectLogoTest.java
@@ -16,6 +16,7 @@
 
 package com.example.video;
 
+import video.LogoDetection;
 import static com.google.common.truth.Truth.assertThat;
 
 import java.io.ByteArrayOutputStream;
