+++ samples/snippets/src/test/java/com/example/video/DetectTextTest.java
@@ -17,6 +17,7 @@
 
 package com.example.video;
 
+import video.TextDetection;
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.cloud.videointelligence.v1.TextAnnotation;
